requirements.txt
rdkit==2025.9.1
numpy==2.2.6
pandas==2.3.3
tqdm==4.67.1
scipy==1.15.3
google_api_python_client==2.185.0
google_auth==2.41.1